Permit overwrite of MdxNode field addition to page context
The block of code doesn't work in conjunction with some customization (e.g. when `page.component` is defined to point to a template or when pages are not in `src/pages`. Adding a blog to avoid re-running the block of code if the MdxNode has been already defined will permit to create additional `onCreatePage` method to apply custom logic.